refactor(ApplyFor): is separated into form components in ApplyFor for readability
